Ignore duplicate opt-in submissions (see #7340)
Description
-----------

Yesterday I was doing some front end tests with a client of mine. They use Microsoft Outlook and some corporate servers for their emails. When they click a link in an email, Outlook (or whatever they use) does a security-check on the URL, which actually queries the URL in the background, before opening the browser tab. This causes problems with (some of) our opt-in tokens. In my case this was about the user registration, because they will receive a _token already confirmed_ error message.

I remember we discussed that in a public call a while ago. Dunno what solution we came up with (except for using POST requets, which isn't easy). The most simple solution I can think of (and this PR implements) is to simply ignore a second submission of a valid token (= within 24hrs). It now shows the same confirmation page or message, but simply does not execute the confirmation action twice.

PS: this does not apply to `ModulePassword`, because there the opt-in token is only confirmed after the form has been manually submitted (with a new password).

Commits
-------

d7ad878a Ignore duplicate opt-in submission